pyproject.toml
[tool.poetry]
name = "oasis"
version = "0.1.0"
description = "Job search & application tailor CLI"
authors = ["Kevin Nebel <kevin.nebel@yahoo.com>"]
packages = [{ include = "oasis", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
typer = ">=0.12,<0.13"
click = "^8.1"
rich = "^13"
textual = "^0.80"
python-jobspy = "^1.1"
python-docx = "^1.1"
pdf2docx = "^0.5"
httpx = "^0.27"
beautifulsoup4 = "^4.12"
anthropic = "^0.40"
keyring = "^25"
pydantic = "^2"
pydantic-settings = "^2"

[tool.poetry.group.dev.dependencies]
pyinstaller = { version = "^6", python = "<3.15" }

[tool.poetry.scripts]
oasis = "oasis.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
